further refactor server
