+++ src/deepsparse/server/server.py
@@ -14,27 +14,19 @@
 
 import logging
 import os
+from abc import abstractmethod
 from collections import Counter
 from concurrent.futures import ThreadPoolExecutor
 from copy import deepcopy
 from typing import List
-from abc import abstractmethod
 
 import yaml
 from pydantic import BaseModel
 
 import uvicorn
 from deepsparse.engine import Context
-from deepsparse.loggers import BaseLogger
 from deepsparse.pipeline import Pipeline
-from deepsparse.server.config import (
-    INTEGRATION_LOCAL,
-    INTEGRATION_SAGEMAKER,
-    INTEGRATIONS,
-    EndpointConfig,
-    ServerConfig,
-    SystemLoggingConfig,
-)
+from deepsparse.server.config import ServerConfig, SystemLoggingConfig
 from deepsparse.server.config_hot_reloading import start_config_watcher
 from deepsparse.server.helpers import (
     prep_outputs_for_serialization,
@@ -44,7 +36,7 @@ from deepsparse.server.system_logging import (
     SystemLoggingMiddleware,
     log_system_information,
 )
-from fastapi import FastAPI, UploadFile
+from fastapi import FastAPI, Request, UploadFile
 from fastapi.exceptions import HTTPException
 from starlette.responses import RedirectResponse
 
@@ -60,6 +52,11 @@ class ModelMetaData(BaseModel):
     model_path: str
 
 
+class ProxyPipeline:
+    def __init__(self, pipeline: Pipeline):
+        self.pipeline = pipeline
+
+
 class Server:
     def __init__(self, config_path: str):
         _LOGGER.info(f"config_path: {config_path}")
@@ -67,6 +64,7 @@ class Server:
         with open(config_path) as fp:
             obj = yaml.safe_load(fp)
 
+        self.config_path = config_path
         self.server_config = ServerConfig(**obj)
         _LOGGER.info(f"Using config: {repr(self.server_config)}")
 
@@ -100,9 +98,9 @@ class Server:
         }
 
         if hot_reload_config:
-            _LOGGER.info(f"Watching {config_path} for changes.")
+            _LOGGER.info(f"Watching {self.config_path} for changes.")
             _ = start_config_watcher(
-                config_path, f"http://{host}:{port}/endpoints", 0.5
+                self.config_path, f"http://{host}:{port}/endpoints", 0.5
             )
 
         app = self._build_app()
@@ -121,7 +119,8 @@ class Server:
         route_counts = Counter([cfg.route for cfg in self.server_config.endpoints])
         if route_counts[None] > 1:
             raise ValueError(
-                "You must specify `route` for all endpoints if multiple endpoints are used."
+                "You must specify `route` for all endpoints if multiple endpoints "
+                "are used."
             )
 
         for route, count in route_counts.items():
@@ -137,7 +136,6 @@ class Server:
         return self._add_routes(app)
 
     def _base_routes(self) -> FastAPI:
-        
         app = FastAPI()
         app.add_middleware(
             SystemLoggingMiddleware,
@@ -200,232 +198,35 @@ class Server:
     def _add_routes(self, app):
         pass
 
-class DeepsparseServer(Server):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    def _add_routes(self, app):
-        # create pipelines & endpoints
-        @app.post("/endpoints", tags=["endpoints"], response_model=bool)
-        def _add_endpoint_endpoint(cfg: EndpointConfig):
-            if cfg.name is None:
-                cfg.name = f"endpoint-{len(app.routes)}"
-            self._add_endpoint(
-                app,
-                cfg,
-            )
-            # force regeneration of the docs
-            app.openapi_schema = None
-            return True
-
-        @app.delete("/endpoints", tags=["endpoints"], response_model=bool)
-        def _delete_endpoint(cfg: EndpointConfig):
-            _LOGGER.info(f"Deleting endpoint for {cfg}")
-            matching = [r for r in app.routes if r.path == cfg.route]
-            assert len(matching) == 1
-            app.routes.remove(matching[0])
-            # force regeneration of the docs
-            app.openapi_schema = None
-            return True
-
-        for endpoint_config in self.server_config.endpoints:
-            self._add_endpoint(
-                app,
-                endpoint_config,
-            )
-
-        _LOGGER.info(f"Added endpoints: {[route.path for route in app.routes]}")
-        return app
-
-    def _add_endpoint(
-        self,
-        app: FastAPI,
-        endpoint_config: EndpointConfig,
-    ):
-        pipeline_config = endpoint_config.to_pipeline_config()
-        pipeline_config.kwargs["executor"] = self.executor
-
-        _LOGGER.info(f"Initializing pipeline for '{endpoint_config.name}'")
-        pipeline = Pipeline.from_config(pipeline_config, self.context, self.server_logger)
-
-        _LOGGER.info(f"Adding endpoints for '{endpoint_config.name}'")
-        self._add_inference_endpoints(
-            app,
-            endpoint_config,
-            self.server_config.system_logging,
-            pipeline,
-            self.server_config.integration,
-        )
-        self._add_status_and_metadata_endpoints(
-            app, endpoint_config, pipeline, self.server_config.integration
-        )
-
-    def _update_routes(self, app, routes_and_fns, response_model, methods, tags):
-        for route, endpoint_fn in routes_and_fns:
-            app.add_api_route(
-                route,
-                endpoint_fn,
-                response_model=response_model,
-                methods=methods,
-                tags=tags,
-            )
-            _LOGGER.info(f"Added '{route}' endpoint")
-
-    def clean_up_route(self, route):
-        if not route.startswith("/"):
-            route = "/" + route
-        return route
-    
-    def _pipeline_ready(self):
+    @staticmethod
+    def pipeline_ready():
         return CheckReady(status="OK")
 
-    def _add_status_and_metadata_endpoints(
-        self,
-        app: FastAPI,
-        endpoint_config: EndpointConfig,
-        pipeline: Pipeline,
-        integration: str = INTEGRATION_LOCAL,
-    ):
-        routes_and_fns = []
-        meta_and_fns = []
+    @staticmethod
+    def model_metadata(proxy_pipeline: ProxyPipeline):
+        if not proxy_pipeline or not proxy_pipeline.pipeline.model_path:
+            HTTPException(status_code=404, detail="Model path not found")
+        return ModelMetaData(model_path=proxy_pipeline.pipeline.model_path)
 
-        def _model_metadata(pipeline):
-            if not pipeline or not pipeline.model_path:
-                HTTPException(status_code=404, detail="Model path not found")
-            return ModelMetaData(model_path=pipeline.model_path)
-
-        if endpoint_config.route:
-            endpoint_config.route = self.clean_up_route(endpoint_config.route)
-            route_ready = f"{endpoint_config.route}/ready"
-            route_meta = endpoint_config.route
-        else:
-            route_ready = f"/v2/models/{endpoint_config.name}/ready"
-            route_meta = f"/v2/models/{endpoint_config.name}"
-
-        routes_and_fns.append((route_ready, self._pipeline_ready))
-        meta_and_fns.append((route_meta, _model_metadata))
-
-        self._update_routes(
-            app, meta_and_fns, ModelMetaData, ["GET"], ["model", "metadata"]
-        )
-        self._update_routes(
-            app, routes_and_fns, CheckReady, ["GET"], ["model", "health"]
-        )
-
-    def _add_inference_endpoints(
-        self,
-        app: FastAPI,
-        endpoint_config: EndpointConfig,
+    @staticmethod
+    async def predict(
+        proxy_pipeline: ProxyPipeline,
         system_logging_config: SystemLoggingConfig,
-        pipeline: Pipeline,
-        integration: str = INTEGRATION_LOCAL,
+        raw_request: Request,
     ):
-        input_schema = pipeline.input_schema
-        output_schema = pipeline.output_schema
-
-        def _predict(request: pipeline.input_schema):
-            pipeline_outputs = pipeline(request)
-            server_logger = pipeline.logger
-            if server_logger:
-                log_system_information(
-                    server_logger=server_logger,
-                    system_logging_config=system_logging_config,
-                )
-            pipeline_outputs = prep_outputs_for_serialization(pipeline_outputs)
-            return pipeline_outputs
-
-        def _predict_from_files(request: List[UploadFile]):
-            request = pipeline.input_schema.from_files(
-                (file.file for file in request), from_server=True
+        request = proxy_pipeline.pipeline.input_schema(**await raw_request.json())
+        pipeline_outputs = proxy_pipeline.pipeline(request)
+        server_logger = proxy_pipeline.pipeline.logger
+        if server_logger:
+            log_system_information(
+                server_logger=server_logger, system_logging_config=system_logging_config
             )
-            return _predict(request)
+        pipeline_outputs = prep_outputs_for_serialization(pipeline_outputs)
+        return pipeline_outputs
 
-        routes_and_fns = []
-        route = (
-            f"{endpoint_config.route}/infer"
-            if endpoint_config.route
-            else f"/v2/models/{endpoint_config.name}/infer"
+    @staticmethod
+    def predict_from_files(proxy_pipeline: ProxyPipeline, request: List[UploadFile]):
+        request = proxy_pipeline.pipeline.input_schema.from_files(
+            (file.file for file in request), from_server=True
         )
-        route = self.clean_up_route(route)
-
-        routes_and_fns.append((route, _predict))
-        if hasattr(input_schema, "from_files"):
-            routes_and_fns.append((route + "/from_files", _predict_from_files))
-
-        self._update_routes(
-            app, routes_and_fns, output_schema, ["POST"], ["model", "inference"]
-        )
-
-class SagemakerServer(DeepsparseServer):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-    
-    def _add_routes(self, app):
-        return super()._add_routes(app)
-
-    def _add_inference_endpoints(
-        self,
-        app: FastAPI,
-        endpoint_config: EndpointConfig,
-        system_logging_config: SystemLoggingConfig,
-        pipeline: Pipeline,
-        integration: str = INTEGRATION_LOCAL,
-    ):
-        input_schema = pipeline.input_schema
-        output_schema = pipeline.output_schema
-
-        def _predict(request: pipeline.input_schema):
-            pipeline_outputs = pipeline(request)
-            server_logger = pipeline.logger
-            if server_logger:
-                log_system_information(
-                    server_logger=server_logger,
-                    system_logging_config=system_logging_config,
-                )
-            pipeline_outputs = prep_outputs_for_serialization(pipeline_outputs)
-            return pipeline_outputs
-
-        def _predict_from_files(request: List[UploadFile]):
-            request = pipeline.input_schema.from_files(
-                (file.file for file in request), from_server=True
-            )
-            return _predict(request)
-
-        routes_and_fns = []
-        route = "/invocations/infer"
-        if hasattr(input_schema, "from_files"):
-            routes_and_fns.append((route, _predict_from_files))
-        else:
-            routes_and_fns.append((route, _predict))
-
-        self._update_routes(
-            app, routes_and_fns, output_schema, ["POST"], ["model", "inference"]
-        )
-
-    def _add_status_and_metadata_endpoints(
-        self,
-        app: FastAPI,
-        endpoint_config: EndpointConfig,
-        pipeline: Pipeline,
-        integration: str = INTEGRATION_LOCAL,
-    ):
-        routes_and_fns = []
-        meta_and_fns = []
-
-        def _model_metadata(pipeline):
-            if not pipeline or not pipeline.model_path:
-                HTTPException(status_code=404, detail="Model path not found")
-            return ModelMetaData(model_path=pipeline.model_path)
-
-        route_ready = "/invocations/ready"
-        route_meta = "/invocations"
-
-        routes_and_fns.append((route_ready, self._pipeline_ready))
-        meta_and_fns.append((route_meta, _model_metadata))
-
-        self._update_routes(
-            app, meta_and_fns, ModelMetaData, ["GET"], ["model", "metadata"]
-        )
-        self._update_routes(
-            app, routes_and_fns, CheckReady, ["GET"], ["model", "health"]
-        )
+        return Server.predict(request)
